+++ Untitled RPG/Assets/Scripts/Items/ItemBase.cs
@@ -7,7 +7,7 @@ public class ItemBase : ScriptableObject
     [SerializeField] private string _description;
     [field: SerializeField, FormerlySerializedAs("_message")] public string Message { get; private set; }
     [field: SerializeField, FormerlySerializedAs("_icon")] public Sprite Icon { get; private set; }
-    [field: SerializeField, FormerlySerializedAs("_price")] public float Price { get; private set; }
+    [field: SerializeField, FormerlySerializedAs("_price")] public int Price { get; private set; }
     [field: SerializeField, FormerlySerializedAs("_isSellable")] public bool IsSellable { get; private set; }
 
     public virtual string Name => _name;
